feat: Replace static articles with dynamic content from Supabase, add featured game filter and author support
